Changed conditions for downshifting
Now if the robot is slowing down, and the operator is applying a lot of
power, it will downshift as soon as it's below the max low gear speed.
If the operator is applying a lot of power and the robot is below the
normal downshift speed threshold, it will also downshift.